+++ packages/viz/__tests__/viewport-features/cases/points/tiles.json
