+++ packages/viz/__tests__/viewport-features/cases/polygons/polygons.ts
@@ -1,7 +1,7 @@
 import { Vector3 } from '@math.gl/core';
-import { ViewportFrustumPlanes } from '../../../geometry/types';
-import { ViewportFeaturesOptions } from '../../../ViewportFeaturesGenerator';
-import { AggregationTypes } from '../../../aggregations';
+import { ViewportFrustumPlanes } from '../../../../src/lib/interactivity/viewport-features/geometry/types';
+import { ViewportFeaturesOptions } from '../../../../src/lib/interactivity/viewport-features/ViewportFeaturesGenerator';
+import { AggregationTypes } from '../../../../src/lib/interactivity/viewport-features/aggregations';
 import * as tiles from './tiles.json';
 
 export {
