+++ packages/viz/jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   preset: 'ts-jest',
-  modulePathIgnorePatterns: ['__tests__/cases/']
+  modulePathIgnorePatterns: ['__tests__/.*/cases/']
 };
